Get rid of missing newline error in clang.
On Darwin, executing g++ runs clang, which emits an error when a C++
code block does not end in a newline. This is different from how g++
behaves on Linux.

This error doesn't add any value in the context of a code block in an
RST file, so we just always add a newline to the C++ source code of a
block to always get rid of it.